+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [unreleased]
+
+### Deprecated
+
+* `SimpleDialog` element now uses `intent` property instead of `intentClassName`
+
 ## [21.11.1] 2021-11-24
 
 ### Added